+++ src/components/Accordion/Accordion.js
@@ -5,20 +5,28 @@ import React, { useState } from "react";
 import AccordionSection from "../AccordionSection";
 
 const generateSections = (sections, setExpanded, expanded) =>
-  sections.map((props, i) => (
+  sections.map(({ key, ...props }, i) => (
     <AccordionSection
       expanded={expanded}
-      key={i}
+      key={key || i}
+      sectionKey={key}
       setExpanded={setExpanded}
       {...props}
     />
   ));
 
-const Accordion = ({ className, onExpandedChange, sections, ...props }) => {
-  const [expanded, setExpanded] = useState();
-  const changeExpanded = (id, title) => {
-    setExpanded(id);
-    onExpandedChange && onExpandedChange(title);
+const Accordion = ({
+  className,
+  expanded,
+  externallyControlled,
+  onExpandedChange,
+  sections,
+  ...props
+}) => {
+  const [expandedSection, setExpandedSection] = useState(expanded);
+  const setExpanded = (id, title) => {
+    setExpandedSection(id);
+    onExpandedChange && onExpandedChange(id, title);
   };
   return (
     <aside
@@ -28,7 +36,11 @@ const Accordion = ({ className, onExpandedChange, sections, ...props }) => {
       aria-multiselectable="true"
     >
       <ul className="p-accordion__list">
-        {generateSections(sections, changeExpanded, expanded)}
+        {generateSections(
+          sections,
+          setExpanded,
+          externallyControlled ? expanded : expandedSection
+        )}
       </ul>
     </aside>
   );
@@ -40,10 +52,29 @@ Accordion.propTypes = {
    */
   className: PropTypes.string,
   /**
+   * An optional value to set the expanded section. The value must match a
+   * section key. This value will only set the expanded section on first render
+   * if externallyControlled is not set to `true`.
+   */
+  expanded: PropTypes.string,
+  /**
+   * Whether the expanded section will be controlled via external state.
+   */
+  externallyControlled: PropTypes.bool,
+  /**
    * An array of sections and content.
    */
   sections: PropTypes.arrayOf(
     PropTypes.shape({
+      /**
+       * The content of the section.
+       */
+      content: PropTypes.node,
+      /**
+       * An optional key for the section component. It will also be
+       * used to track which section is selected.
+       */
+      key: PropTypes.string,
       /**
        * An optional click event when the title is clicked.
        */
@@ -51,11 +82,7 @@ Accordion.propTypes = {
       /**
        * The title of the section.
        */
-      title: PropTypes.string,
-      /**
-       * The content of the section.
-       */
-      content: PropTypes.node
+      title: PropTypes.string
     }).isRequired
   ),
   /**